requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0 
django-filter==25.1